+++ tests/fixtures/projects/react-component-with-css/src/components/Thing.js
@@ -1,0 +1,9 @@
+import './Thing.css'
+
+import React, {Component} from 'react'
+
+export default class Thing extends Component {
+  render() {
+    return <div className="Thing"/>
+  }
+}