+++ tests/fixtures/projects/react-component-with-css/src/index.js
@@ -1,0 +1,1 @@
+export {default as Thing} from './components/Thing'